Added hook for Topic page Outsourced mapLearningPathToNodes function to hook
