+++ src/pages/Topic/Topic.tsx
@@ -11,10 +11,11 @@ import StudentLearningElement from 'src/common/core/StudentLearningElement/Stude
 import { useTranslation } from 'react-i18next'
 import 'reactflow/dist/style.css'
 import { DefaultSkeleton as Skeleton } from '@common/components'
+import { useTopic as _useTopic, TopicHookReturn } from './Topic.hooks'
 
-const _useTopic = () => {
+/*const _useTopic = () => {
   console.log('useTopic')
-}
+}*/
 
 const studentLearningElement: StudentLearningElement = {
   id: 160,
@@ -158,13 +159,13 @@ const learningPath: LearningPath = {
   path: learningPathLearningElements
 }
 
-type TopicProps = {
+export type TopicProps = {
   useTopic?: typeof _useTopic
 }
 
 // TODO URL Stuktur übelrgeen. bzswp. localhost:3000/topic?topic=1
 // Topic Page - TODO Component extract
-export const Topic = ({ useTopic = _useTopic }: TopicProps): JSX.Element => {
+const Topic = ({ useTopic = _useTopic }: TopicProps): JSX.Element => {
   const [initialNodes, setInitialNodes] = useState<Node[]>()
   const [initialEdges, setInitialEdges] = useState<Edge[]>()
   const navigate = useNavigate()
@@ -176,26 +177,32 @@ export const Topic = ({ useTopic = _useTopic }: TopicProps): JSX.Element => {
   const [title, setTitle] = useState('')
   const [isOpen, setIsOpen] = useState(false)
 
+  const { mapLearningPathToNodes } = useTopic()
+
   const fetchUser = useBoundStore((state) => state.fetchUser)
   const fetchLearningPath = useBoundStore((state) => state.fetchLearningPath)
 
-  const handleOpen = useMemo(() => {
-    return () => {
-      setIsOpen(true)
-    }
-  }, [])
+  const handleOpen = useCallback(() => {
+    setIsOpen(true)
+  }, [setIsOpen])
 
-  const handleClose = () => {
+  const handleClose = useCallback(() => {
     setIsOpen(false)
-  }
+  }, [setIsOpen])
 
-  const handleSetUrl = useCallback((url: string) => {
-    setUrl(url)
-  }, [])
+  const handleSetUrl = useCallback(
+    (url: string) => {
+      setUrl(url)
+    },
+    [setUrl]
+  )
 
-  const handleSetTitle = useCallback((title: string) => {
-    setTitle(title)
-  }, [])
+  const handleSetTitle = useCallback(
+    (title: string) => {
+      setTitle(title)
+    },
+    [setTitle]
+  )
 
   const mapNodes = useCallback(
     (learning_path_data: LearningPath) => {
@@ -221,43 +228,34 @@ export const Topic = ({ useTopic = _useTopic }: TopicProps): JSX.Element => {
     [theme]
   )
 
-  // useEffect(() => {
-  //   const { nodes, edges } = mapNodes(learningPath)
-  //   console.log('rendering nodes')
-  //   setInitialNodes(nodes)
-  //   //setInitialEdges(edges)
-  //   setInitialEdges(edges)
-  // }, [mapNodes, learningPath])
-
   useEffect(() => {
     // request to backend to get learning path for topic
     // alert('Topic: ' + topic)
-    log.log('Topic page')
-    const preventEndlessLoading = setTimeout(() => {
-      log.log('Topic page timeout')
-      navigate('/login')
-    }, 5000)
-    if (authcontext.isAuth) {
-      clearTimeout(preventEndlessLoading)
-      fetchUser().then((user) => {
-        fetchLearningPath(user.settings.user_id, user.lms_user_id, user.id, Number(courseId), Number(topicId)).then(
-          (learning_path_data) => {
-            const { nodes, edges } = mapNodes(learning_path_data)
-            setInitialNodes(nodes)
-            //setInitialEdges(edges)
-            setInitialEdges(edges)
-          }
-        )
-      })
-    }
+    // log.log('Topic page')
+    // const preventEndlessLoading = setTimeout(() => {
+    //   log.log('Topic page timeout')
+    //   navigate('/login')
+    // }, 5000)
+    //if (authcontext.isAuth) {
+    //  clearTimeout(preventEndlessLoading)
+    //  fetchUser().then((user) => {
+    // fetchLearningPath(user.settings.user_id, user.lms_user_id, user.id, Number(courseId), Number(topicId)).then(
+    //  (learning_path_data) => {
+    //const { nodes, edges } = mapNodes(learning_path_data)
+    const { nodes, edges } = mapNodes(learningPath)
+    setInitialNodes(nodes)
+    //setInitialEdges(edges)
+    setInitialEdges(edges)
+    //})
+    // })
+    // }
     return () => {
-      clearTimeout(preventEndlessLoading)
+      // clearTimeout(preventEndlessLoading)
     }
   }, [authcontext.isAuth, courseId, fetchLearningPath, fetchUser, theme, topicId])
 
   log.setLevel('error')
 
-  // TODO: HIer edges rendern
   return initialNodes && initialEdges ? (
     <Box height={'100%'}>
       <ReactFlow nodes={initialNodes} edges={initialEdges} nodeTypes={nodeTypes} fitView>
@@ -273,146 +271,3 @@ export const Topic = ({ useTopic = _useTopic }: TopicProps): JSX.Element => {
 }
 
 export default Topic
-
-const mapLearningPathToNodes = (
-  learningPath: LearningPath,
-  theme: Theme,
-  handleSetUrl: (url: string) => void,
-  handleSetTitle: (title: string) => void,
-  handleOpen: () => void,
-  handleClose: () => void
-): Node[] => {
-  // Sort learning path
-  const sortedLearningPath = learningPath.path.sort((a, b) => a.position - b.position)
-
-  // Every exercise learning element
-  const learningPathExercises = sortedLearningPath.filter((item) => item.learning_element.classification === 'ÜB')
-
-  // Every learning element except exercises
-  const learningPathExcludingExercises = sortedLearningPath.filter(
-    (item) => item.learning_element.classification !== 'ÜB'
-  )
-
-  const groupHeight = 175
-  const nodeOffsetX = 50
-
-  const learningElementStyle = {
-    background: theme.palette.primary.main,
-    padding: 10,
-    border: '1px solid ' + theme.palette.grey[500],
-    borderRadius: 8,
-    cursor: 'pointer'
-  }
-
-  // Exercise nodes
-  const exerciseLearningElementChildNodes = learningPathExercises.map((node, index) => {
-    const node_data: LearningPathLearningElementNode = {
-      lms_id: node.learning_element.lms_id,
-      name: node.learning_element.name,
-      activity_type: node.learning_element.activity_type,
-      classification: node.learning_element.classification,
-      is_recommended: node.recommended,
-      handleSetUrl: handleSetUrl,
-      handleSetTitle: handleSetTitle,
-      handleOpen: handleOpen,
-      handleClose: handleClose
-    }
-    return {
-      id: node.position.toString() + '-' + node.learning_element.lms_id,
-      type: node.learning_element.classification,
-      data: node_data,
-      position: {
-        x: nodeOffsetX + 300 * index,
-        y: 250 * (learningPathExercises[0].position - 1) + 50
-      },
-      style: learningElementStyle
-    }
-  })
-
-  // Parent node for exercise learning elements
-  const exerciseLearningElementParentNode =
-    learningPathExercises.length > 0
-      ? {
-        id: learningPathExercises[0].position.toString(),
-        data: { label: 'Übungen' },
-        type: 'GROUP',
-        position: {
-          x: 0,
-          y: 250 * (learningPathExercises[0].position - 1)
-        },
-        style: {
-          border: '1px solid ' + theme.palette.grey[500],
-          borderRadius: 8,
-          width: 300 * learningPathExercises.length + nodeOffsetX,
-          height: groupHeight
-        }
-      }
-      : null
-
-  // Rest of learning elements
-  const learningElementNodesExcludingExercises = learningPathExcludingExercises.map((item, index) => {
-    const node_data: LearningPathLearningElementNode = {
-      lms_id: item.learning_element.lms_id,
-      name: item.learning_element.name,
-      activity_type: item.learning_element.activity_type,
-      classification: item.learning_element.classification,
-      is_recommended: item.recommended,
-      handleSetUrl: handleSetUrl,
-      handleSetTitle: handleSetTitle,
-      handleOpen: handleOpen,
-      handleClose: handleClose
-    }
-
-    return {
-      id: item.position.toString(),
-      type: item.learning_element.classification,
-      data: node_data,
-      position: {
-        x: nodeOffsetX + (300 * (learningPathExercises.length - 1)) / 2,
-        y:
-          exerciseLearningElementParentNode && item.position < parseInt(exerciseLearningElementParentNode.id)
-            ? 250 * (item.position - 1)
-            : exerciseLearningElementParentNode
-              ? 250 * (item.position - exerciseLearningElementChildNodes.length) + groupHeight - 70
-              : 250 * (item.position - 1)
-      },
-      style: learningElementStyle
-    }
-  })
-
-  // Insert exercise nodes into learning elements
-  const learningElementNodesBeforeExercises = learningElementNodesExcludingExercises.filter(
-    (item) => exerciseLearningElementParentNode && parseInt(item.id) < parseInt(exerciseLearningElementParentNode.id)
-  )
-
-  const learningElementNodesAfterExercises = learningElementNodesExcludingExercises.filter(
-    (item) => exerciseLearningElementParentNode && parseInt(item.id) > parseInt(exerciseLearningElementParentNode.id)
-  )
-
-  // Add 1 to the id of all elements (including exercises) after the exercise parent node
-  const learningElementNodesAfterExercisesElementParentNode = learningElementNodesAfterExercises.map((item) => {
-    return {
-      ...item,
-      id: (parseInt(item.id) + 1).toString()
-    }
-  })
-
-  // const exerciseLearningElementChildNodesWithIncreasedId = exerciseLearningElementChildNodes.map((item) => {
-  //   return {
-  //     ...item,
-  //     id: (parseInt(item.id) + 1).toString()
-  //   }
-  // })
-
-  const learningElementNodes = [
-    ...(learningElementNodesBeforeExercises.length > 0
-      ? learningElementNodesBeforeExercises
-      : learningElementNodesExcludingExercises),
-    ...(exerciseLearningElementParentNode
-      ? [exerciseLearningElementParentNode, ...exerciseLearningElementChildNodes]
-      : []),
-    ...learningElementNodesAfterExercisesElementParentNode
-  ]
-
-  return learningElementNodes
-}
